[refactor] Ensure registered module be subclass of Module
